在庫編集 useState -> redux置き換え
