+++ client/src/components/Stock/index.tsx
@@ -28,6 +28,8 @@ export type TStorehouse = {
 
 const store = configureStore({ reducer: rootReducer })
 
+export type RootState = ReturnType<typeof store.getState>
+
 const Stock = () => {
   return (
     <Provider store={store}>
